Fix plot_tomography_l1_reconstruction.py
Error was: TypeError: 'numpy.float64' object cannot be interpreted as an integer.

[doc build]

+++ examples/applications/plot_tomography_l1_reconstruction.py
@@ -51,7 +51,7 @@ import matplotlib.pyplot as plt
 
 def _weights(x, dx=1, orig=0):
     x = np.ravel(x)
-    floor_x = np.floor((x - orig) / dx)
+    floor_x = np.floor((x - orig) / dx).astype(np.int64)
     alpha = (x - orig - floor_x * dx) / dx
     return np.hstack((floor_x, floor_x + 1)), np.hstack((1 - alpha, alpha))
